# services/api-gateway/docs/swagger.yaml
basePath: /api/v1
definitions:
  dto.CreateOfficerRequest:
    properties:
      email:
        type: string
      password:
        maxLength: 72
        minLength: 8
        type: string
    required:
    - email
    - password
    type: object
  dto.Customer:
    properties:
      created_at:
        type: string
      id:
        type: string
      profile:
        $ref: '#/definitions/dto.Profile'
      status:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  dto.CustomerStatus:
    properties:
      changed_at:
        type: string
      status:
        type: string
    type: object
  dto.ErrorResponse:
    properties:
      message:
        type: string
    type: object
  dto.ListCustomers:
    properties:
      customers:
        items:
          $ref: '#/definitions/dto.Customer'
        type: array
    type: object
  dto.LoginRequest:
    properties:
      email:
        type: string
      password:
        maxLength: 72
        minLength: 8
        type: string
    required:
    - email
    - password
    type: object
  dto.LoginResponse:
    properties:
      tokens:
        $ref: '#/definitions/dto.TokenPair'
      user:
        $ref: '#/definitions/dto.User'
    type: object
  dto.LogoutRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  dto.Profile:
    properties:
      birth_date:
        type: string
      citizenship:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      phone:
        type: string
    type: object
  dto.RefreshRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  dto.RegisterRequest:
    properties:
      email:
        type: string
      password:
        maxLength: 72
        minLength: 8
        type: string
    required:
    - email
    - password
    type: object
  dto.RegisterResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      role:
        type: string
      user_id:
        type: string
    type: object
  dto.TokenPair:
    properties:
      access_expires_at:
        type: string
      access_token:
        type: string
      refresh_expires_at:
        type: string
      refresh_token:
        type: string
    type: object
  dto.UpdateProfile:
    properties:
      birth_date:
        type: string
      citizenship:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      phone:
        type: string
    type: object
  dto.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      role:
        type: string
    type: object
  dto.ValidateTokenRequest:
    properties:
      access_token:
        type: string
    required:
    - access_token
    type: object
  dto.ValidateTokenResponse:
    properties:
      expires_at:
        type: string
      id:
        type: string
      role:
        type: string
      valid:
        type: boolean
    type: object
info:
  contact:
    name: Danila Kuryakin
    url: https://github.com/danila-kuryakin/banking_mini_cores
  description: 'HTTP-фасад над gRPC-сервисами: аутентификация и работа с клиентами.'
  title: Banking Mini Cores API Gateway
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Возвращает данные пользователя и пару токенов. Access-токен передаётся
        дальше в заголовке Authorization.
      parameters:
      - description: Email и пароль
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Пользователь и пара токенов
          schema:
            $ref: '#/definitions/dto.LoginResponse'
        "400":
          description: Тело запроса не прошло валидацию
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Неверные учётные данные
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Ошибка auth-service
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Вход по email и паролю
      tags:
      - auth
  /auth/logout:
    post:
      consumes:
      - application/json
      description: Отзывает refresh-токен. Ранее выданный access-токен продолжает
        действовать до истечения срока.
      parameters:
      - description: Refresh-токен
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.LogoutRequest'
      produces:
      - application/json
      responses:
        "204":
          description: Сессия завершена, тело ответа пустое
        "400":
          description: Тело запроса не прошло валидацию или токен неизвестен
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Ошибка auth-service
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Выход из сессии
      tags:
      - auth
  /auth/officers:
    post:
      consumes:
      - application/json
      description: Заводит учётную запись с ролью officer. Доступно только роли admin.
      parameters:
      - description: Email и пароль сотрудника
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateOfficerRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Сотрудник создан
          schema:
            $ref: '#/definitions/dto.User'
        "400":
          description: Тело запроса не прошло валидацию
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Отсутствует или недействителен access-токен
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Роль не admin
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "409":
          description: Email уже зарегистрирован
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Ошибка auth-service
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Создание сотрудника
      tags:
      - auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Выдаёт новую пару токенов по refresh-токену. Старый refresh-токен
        после обмена недействителен.
      parameters:
      - description: Refresh-токен
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Новая пара токенов
          schema:
            $ref: '#/definitions/dto.TokenPair'
        "400":
          description: Тело запроса не прошло валидацию
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Refresh-токен просрочен или отозван
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Ошибка auth-service
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Обновление пары токенов
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Создаёт учётную запись клиента. Пароль — от 8 до 72 символов.
      parameters:
      - description: Email и пароль
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Пользователь создан
          schema:
            $ref: '#/definitions/dto.RegisterResponse'
        "400":
          description: Тело запроса не прошло валидацию
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "409":
          description: Email уже зарегистрирован
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Ошибка auth-service
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Регистрация пользователя
      tags:
      - auth
  /auth/validate:
    post:
      consumes:
      - application/json
      description: 'Служебный метод: сообщает, валиден ли токен, и возвращает идентификатор
        владельца, роль и срок действия.'
      parameters:
      - description: Access-токен
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.ValidateTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Результат проверки; valid=false для просроченного токена
          schema:
            $ref: '#/definitions/dto.ValidateTokenResponse'
        "400":
          description: Тело запроса не прошло валидацию
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Ошибка auth-service
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Проверка access-токена
      tags:
      - auth
  /customer:
    get:
      description: Постраничный список клиентов, отсортированный по дате создания
        по убыванию. Доступно ролям officer и admin.
      parameters:
      - description: Размер страницы, по умолчанию 20
        in: query
        maximum: 100
        minimum: 0
        name: limit
        type: integer
      - description: Сколько записей пропустить
        in: query
        minimum: 0
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Страница списка
          schema:
            $ref: '#/definitions/dto.ListCustomers'
        "400":
          description: Некорректные параметры запроса
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Отсутствует или недействителен access-токен
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Роль не officer и не admin
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Список клиентов
      tags:
      - customer
  /customer/{user_id}:
    get:
      description: Возвращает клиента вместе с профилем и текущим статусом. Клиент
        видит только себя, officer и admin - любого.
      parameters:
      - description: UUID пользователя или me - текущий пользователь
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Клиент найден
          schema:
            $ref: '#/definitions/dto.Customer'
        "401":
          description: Отсутствует или недействителен access-токен
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Профиль принадлежит другому пользователю
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Клиент не найден
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Карточка клиента
      tags:
      - customer
    post:
      consumes:
      - application/json
      description: Записывает анкетные данные. Поле birth_date передаётся строкой
        в формате YYYY-MM-DD и может быть пустым. Когда профиль заполнен целиком,
        клиент переходит в статус profile_filled и дальнейшее редактирование запрещается.
      parameters:
      - description: UUID пользователя или me - текущий пользователь
        in: path
        name: user_id
        required: true
        type: string
      - description: Поля профиля
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateProfile'
      produces:
      - application/json
      responses:
        "200":
          description: Профиль обновлён
          schema:
            $ref: '#/definitions/dto.Customer'
        "400":
          description: Некорректное тело запроса или профиль больше не редактируется
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Отсутствует или недействителен access-токен
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Профиль принадлежит другому пользователю
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Клиент не найден
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Обновление профиля клиента
      tags:
      - customer
  /customer/{user_id}/status:
    get:
      description: 'Возвращает текущий статус и момент его последней смены. Возможные
        значения: new, profile_filled, on_kyc, active, rejected, blocked.'
      parameters:
      - description: UUID пользователя или me - текущий пользователь
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Текущий статус
          schema:
            $ref: '#/definitions/dto.CustomerStatus'
        "401":
          description: Отсутствует или недействителен access-токен
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Профиль принадлежит другому пользователю
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Клиент не найден
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Статус клиента
      tags:
      - customer
securityDefinitions:
  BearerAuth:
    description: Access-токен из POST /auth/login в формате "Bearer <token>".
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
tags:
- description: Регистрация, вход, обновление и отзыв токенов
  name: auth
- description: Карточка клиента, профиль и статус
  name: customer